feat: add common.js for importer whith *
